// src/integrations/supabase/types.ts
export type EstatisticaOportunidadePorEmpresa = {
  empresa: string;
  enviadas: number;
  recebidas: number;
  saldo: number;
  por_status: {
    status: string;
    quantidade: number;
  }[];
};

export type EstatisticaParceiroExterno = {
  parceiro: string;
  enviadas: number;
  recebidas: number;
  saldo: number;
};